+++ api/urls.py
@@ -50,6 +50,7 @@ from api.views import (
     PurchaseOptionsView,
     PurchaseRetrieveUpdateDestroyView,
     PurchasesDeleteView,
+    PurchasesImportOldView,
     PurchasesImportView,
     PurchasesRestoreView,
     RemoveManagerView,
@@ -183,6 +184,7 @@ urlpatterns = {
         name="remove_manager",
     ),
     path("importPurchases/", PurchasesImportView.as_view(), name="purchases_import"),
+    path("importPurchasesOld/", PurchasesImportOldView.as_view(), name="purchases_import_old"),
     path("importCanteens/create/", CanteensCreateImportView.as_view(), name="canteens_create_import"),
     path("importCanteens/update/", CanteensUpdateImportView.as_view(), name="canteens_update_import"),
     path("importCanteensManagers/", CanteensManagersImportView.as_view(), name="canteens_managers_import"),
