+++ api/views/__init__.py
@@ -60,6 +60,7 @@ from .purchase import (  # noqa: F401
     PurchasesRestoreView,
 )
 from .purchase_import import PurchasesImportView  # noqa: F401
+from .purchase_import_old import PurchasesImportOldView  # noqa: F401
 from .reservationexpe import ReservationExpeView  # noqa: F401
 from .resourceaction import ResourceActionView  # noqa: F401
 from .review import ReviewView  # noqa: F401
